+++ CHANGES.md
@@ -1,3 +1,7 @@
+# [0.12.0](https://github.com/ComplianceAsCode/auditree-arboretum/releases/tag/v0.12.0)
+
+- [ADDED] Github org integrity teams fetcher functionality added to Github org integrity permissions fetcher.
+
 # [0.11.1](https://github.com/ComplianceAsCode/auditree-arboretum/releases/tag/v0.11.1)
 
 - [FIXED] Renamed RepoMetadataEvidence `filtered_content` to `relevant_content`.
@@ -5,7 +9,7 @@
 # [0.11.0](https://github.com/ComplianceAsCode/auditree-arboretum/releases/tag/v0.11.0)
 
 - [ADDED] Github org integrity fetcher added to `permissions`.
-- [ADDED] Github org integrity forks fetcher functionality added Github org integrity fetcher.
+- [ADDED] Github org integrity forks fetcher functionality added to Github org integrity permissions fetcher.
 - [CHANGED] Github org integrity collaborators fetcher functionality added to Github org integrity permissions fetcher.
 
 # [0.10.0](https://github.com/ComplianceAsCode/auditree-arboretum/releases/tag/v0.10.0)